Refactor JSJavaBridge, tool window and project utils
