+++ src/main/kotlin/ai/devchat/core/handlers/AddContextRequestHandler.kt
@@ -1,8 +1,8 @@
 package ai.devchat.core.handlers
 
-import ai.devchat.common.ProjectUtils
 import ai.devchat.core.BaseActionHandler
 import ai.devchat.core.DevChatActions
+import ai.devchat.plugin.currentProject
 import com.alibaba.fastjson.JSONObject
 import java.io.BufferedReader
 import java.io.File
@@ -16,7 +16,7 @@ class AddContextRequestHandler(requestAction: String, metadata: JSONObject?, pay
     val command: String? = payload?.getString("command")
 
     override fun action() {
-        val projectDir = ProjectUtils.project?.basePath
+        val projectDir = currentProject?.basePath
         val process = Runtime.getRuntime().exec(command, null, projectDir?.let { File(it) })
         val result = process.inputStream.bufferedReader().use(BufferedReader::readText)
         val errors = process.errorStream.bufferedReader().use(BufferedReader::readText)
