+++ src/main/kotlin/ai/devchat/core/handlers/CommitCodeRequestHandler.kt
@@ -1,9 +1,9 @@
 package ai.devchat.core.handlers
 
-import ai.devchat.common.ProjectUtils
 import ai.devchat.common.Log
 import ai.devchat.core.BaseActionHandler
 import ai.devchat.core.DevChatActions
+import ai.devchat.plugin.currentProject
 import com.alibaba.fastjson.JSONObject
 import java.io.BufferedReader
 import java.io.File
@@ -18,7 +18,7 @@ class CommitCodeRequestHandler(requestAction: String, metadata: JSONObject?, pay
     override fun action() {
         val message = payload!!.getString("message")
         val command = arrayOf("git", "commit", "-m", message)
-        val projectDir = ProjectUtils.project?.basePath
+        val projectDir = currentProject?.basePath
         val process = Runtime.getRuntime().exec(command, null, projectDir?.let { File(it) })
         val result = process.inputStream.bufferedReader().use(BufferedReader::readText)
         val errors = process.errorStream.bufferedReader().use(BufferedReader::readText)
