+++ src/main/kotlin/ai/devchat/core/handlers/ReplaceFileContentHandler.kt
@@ -1,8 +1,8 @@
 package ai.devchat.core.handlers
 
-import ai.devchat.common.ProjectUtils
 import ai.devchat.core.BaseActionHandler
 import ai.devchat.core.DevChatActions
+import ai.devchat.plugin.currentProject
 import com.alibaba.fastjson.JSONObject
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.CommandProcessor
@@ -15,12 +15,11 @@ class ReplaceFileContentHandler(requestAction: String, metadata: JSONObject?, pa
 ) {
     override val actionName: String = DevChatActions.REPLACE_FILE_CONTENT_RESPONSE
     override fun action() {
-        val project = ProjectUtils.project
         val newFileContent = payload!!.getString("content")
         ApplicationManager.getApplication().invokeLater {
-            val editor = FileEditorManager.getInstance(project!!).selectedTextEditor
+            val editor = FileEditorManager.getInstance(currentProject!!).selectedTextEditor
             val document = editor!!.document
-            CommandProcessor.getInstance().executeCommand(project, {
+            CommandProcessor.getInstance().executeCommand(currentProject, {
                 ApplicationManager.getApplication().runWriteAction {
                     document.setText(newFileContent)
                 } }, "ReplaceFileContentHandler", null)
