+++ src/main/kotlin/ai/devchat/core/handlers/ShowSettingDialogRequestHandler.kt
@@ -1,8 +1,8 @@
 package ai.devchat.core.handlers
 
-import ai.devchat.common.ProjectUtils
 import ai.devchat.core.BaseActionHandler
 import ai.devchat.core.DevChatActions
+import ai.devchat.plugin.currentProject
 import com.alibaba.fastjson.JSONObject
 import com.intellij.openapi.options.ShowSettingsUtil
 
@@ -14,6 +14,6 @@ class ShowSettingDialogRequestHandler(requestAction: String, metadata: JSONObjec
 ) {
     override val actionName: String = DevChatActions.SHOW_SETTING_DIALOG_REQUEST
     override fun action() {
-        ShowSettingsUtil.getInstance().showSettingsDialog(ProjectUtils.project, "DevChat")
+        ShowSettingsUtil.getInstance().showSettingsDialog(currentProject, "DevChat")
     }
 }
