+++ src/main/kotlin/ai/devchat/plugin/actions/AddToDevChatFileAction.kt
@@ -1,6 +1,6 @@
 package ai.devchat.plugin.actions
 
-import ai.devchat.plugin.DevChatToolWindow
+import ai.devchat.plugin.loaded
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -36,7 +36,7 @@ class AddToDevChatFileAction : AnAction() {
                 ToolWindowManager.getInstance(project).getToolWindow("DevChat")?.show {
                     val bytes = virtualFile.contentsToByteArray()
                     val content = String(bytes, StandardCharsets.UTF_8)
-                    addToDevChatAction.execute(relativePath, content, language, 0, !DevChatToolWindow.loaded)
+                    addToDevChatAction.execute(relativePath, content, language, 0, !loaded)
                 }
             } catch (ex: IOException) {
                 ex.printStackTrace()
